refactor: useSelectedCharacterStore의 패키지 이동
