+++ fourtoon-cookie/src/hooks/store/selectedCharacter.ts
@@ -1,6 +1,6 @@
 import { create } from "zustand";
 import { persist, createJSONStorage } from 'zustand/middleware'
-import { Character } from "../types/character";
+import { Character } from "../../types/character";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 
 interface SelectedCharacterState {
